refactor(chip): apply theme
