+++ src/components/Chip/index.tsx
@@ -1,0 +1,13 @@
+import Styled from './styles';
+import { ChipProps } from './types';
+
+const Chip = ({
+    label = 'Chip',
+    ...props
+}: ChipProps) => (
+    <Styled {...props}>
+        {label}
+    </Styled>
+);
+
+export { Chip };
