+++ src/components/Chip/types.ts
@@ -1,0 +1,6 @@
+import { HTMLAttributes } from 'react';
+
+export interface ChipProps extends HTMLAttributes<HTMLDivElement> {
+    label: string;
+    color: 'grey' | 'blue' | 'green' | 'orange' | 'red';
+}
