+++ src/index.ts
@@ -4,7 +4,7 @@ export * from './components/Button';
 export * from './components/TextInput';
 export * from './components/NumberInput/NumberInput';
 export * from './components/Checkbox';
-export * from './components/Chip/Chip';
+export * from './components/Chip';
 export * from './components/Select/Select';
 export * from './components/Card';
 export * from './components/Radio/Radio';
